requirements.txt
requests~=2.31.0
Unidecode~=1.3.7
python-dotenv~=1.0.0
aiohttp~=3.8.5
shazamio~=0.3.1.1
spotipy~=2.23.0
eyed3~=0.9.7
musicbrainzngs~=0.7.1
telebot~=0.0.5
click~=8.1.3
pyacoustid~=1.3.0
uvloop~=0.18.0
yt_dlp
google-api-python-client
click==8.1.3
shazamio==0.3.1.1
music-tag==0.4.3
python3-discogs-client==2.6
python-magic==0.4.26